backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
python-jose[cryptography]==3.3.0
anthropic==0.34.2
httpx==0.27.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
